+++ CHANGES.rst
@@ -6,3 +6,6 @@ Changelog of sufriblib
 ----------------
 
 - Initial project structure created with nensskel 1.34.dev0.
+
+- Initial code that can parse RIB/RMB files a little and give some
+  error messages.